requirements.txt
asgiref==3.2.10
bower==0.0.0
bower.py==1.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
dj-database-url==0.5.0
Django==3.1
django-extensions==3.0.5
django-filter==2.3.0
django-heroku==0.3.1
django-reversion==3.0.7
djangorestframework==3.11.1
github3.py==1.3.0
gunicorn==20.0.4
idna==2.10
jwcrypto==0.8
Markdown==3.2.2
psycopg2==2.8.5
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
semantic-version==2.8.5
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
